pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-openterms"
version = "0.1.0"
description = "OpenTerms protocol integration for LangChain: permission-aware AI agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "John Stibal"}
]
keywords = ["langchain", "openterms", "ai-agents", "compliance", "permissions"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "langchain-core>=0.2.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
sdk = [
    "openterms-py>=0.1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "responses>=0.23",
    "openterms-py>=0.1.0",
]

[project.urls]
Homepage = "https://openterms.com"
Repository = "https://github.com/jstibal/langchain-openterms"
Documentation = "https://openterms.com/docs"

[tool.setuptools.packages.find]
include = ["langchain_openterms*"]
